requirements.txt
python-telegram-bot
boto3
python-dotenv
schedule
python-telegram-bot[job-queue]

